[Modify] - 북마크 페이지 및 카테고리 페이지에서 사용되는 infinity scroll 코드 커스텀 훅으로 분리
