+++ src/pages/category/CategoryPage.tsx
@@ -1,29 +1,19 @@
 import { db } from '@/api/pocketbase';
 import { Header, LargeCard } from '@/components';
-
 import getPbImage from '@/util/data/getPBImage';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { useInView } from 'react-intersection-observer';
-import { RecordModel } from 'pocketbase';
-import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
-
-
-
+import {DefaultLoader} from '@/components';
+import { useInifinityCard } from '@/hooks/useInfinityCard';
 
 export function CategoryPage() {
   const {title} = useParams();
-  const { ref, inView } = useInView({ threshold: 0.7 });
-  const [userData, setUserData] = useState<RecordModel>();
-
-  const getRecipeData = async ({ pageParam = 1 }) => {
-    
+  
+  async function getRecipeData({pageParam = 1}) {
     if(title === "오늘의 레시피") {
       const recordsData = await db.collection('recipes').getList(pageParam, 6, { 
         expand: 'rating, profile', 
         sort: '-views' });
-
       return recordsData.items;
     } else {
       const recordsData = await db.collection('recipes').getList(pageParam, 6, { 
@@ -31,51 +21,11 @@ export function CategoryPage() {
         filter: `category = "${title}"`,
         sort: '-created'
       });
-
       return recordsData.items
     }
-  };
+  }
 
-  const { data, status, isFetchingNextPage, fetchNextPage, hasNextPage } = useInfiniteQuery({
-    queryKey: ['recipes'],
-    queryFn: getRecipeData,
-    initialPageParam: 1,
-    getNextPageParam: (lastPage, allPages) => {
-      const nextPage = lastPage.length ? allPages.length + 1 : undefined;
-      return nextPage;
-    },
-  });
-
-  useEffect(() => {
-    console.log(data);
-  }, [data]);
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      console.log('isInview');
-      fetchNextPage();
-    }
-  }, [fetchNextPage, hasNextPage, inView]);
-
-  useEffect(() => {
-    async function getUserData() {
-      const currentUser = localStorage.getItem('pocketbase_auth');
-      if (currentUser === null) return;
-      const userId = JSON.parse(currentUser).model.id;
-      const response = await db.collection('users').getOne(userId, { requestKey: null });
-      if (response === undefined) return;
-      setUserData(response);
-    }
-
-    getUserData();
-    db.collection('users').subscribe('*', getUserData);
-
-    return () => {
-      db.collection('users').unsubscribe();
-    };
-  }, []);
-
-
+  const { data, status, isFetchingNextPage, userData, ref } = useInifinityCard(getRecipeData);
   const contents = data?.pages.map((recipes) =>
     recipes.map((recipe, index) => {
       const url = getPbImage('recipes', recipe.id, recipe.image);
@@ -110,8 +60,8 @@ export function CategoryPage() {
     })
   );
 
-  if (status === 'pending') return <div>로딩중~~</div>;
-  if (status === 'error') return <div>실패 ㅋㅋ</div>;
+  if (status === 'pending') return <DefaultLoader />;
+  if (status === 'error') return <DefaultLoader />;
 
   return (
     <div className="w-full h-full bg-gray-200 overflow-auto">
@@ -120,7 +70,7 @@ export function CategoryPage() {
       </Helmet>
       <Header option="titleWithBack" title={title}/>
       <div className="grid gap-6pxr pb-140pxr grid-cols-card justify-center w-full">{contents}</div>
-      {isFetchingNextPage && <p>Loading...</p>}
+      {isFetchingNextPage && <p className='mx-auto w-full'>로딩중</p>}
     </div>
   );
 }
